Fix English tweets being excluded from stream
Previously, the only English tweets that showed up in columns were ones
that had Japanese text in them (typically the name of the boss).

For some reason, filtering the stream by "I need backup!" only returns
tweets if there is text before "I need backup!" (not sure if the text
needs to be Japanese or if it's any text). Twitter's streaming API is
weird when dealing with CJK text and filtering by full words.

This changes it to filter by "I need backup!Battle ID:" instead.

Fixes #59