RaidFinder#getRaidTweets now returns generic type
Previously it would return `RaidTweet` which is usually ok, but once
it reaches the server side, the `RaidTweet` would get converted to a
protobuf message for every client.

Changing it to a generic type allows the protobuf message to be saved in
the cached observable, which greatly reduces the amount of duplicate
conversion/serialization.