Allow mapping the value before caching observable
This allows the key to be derived from a different source than the
output value. E.g., we want to convert a `RaidTweet` to an `Array[Byte]`
and cache that, but there's no easy way to get the boss name from a byte
array. The key deriving function now acts on `RaidTweet` instead of the
`Array[Byte]`.