Optimize serializing outbound Protobuf messages (#115)
The previous code kept the messages as ScalaPB case classes until it was
time to send them to the client, meaning each raid tweet is serialized
into a `Byte[Array]` once per user. This changes it to only serialize a
raid tweet once, as it comes in from the Twitter stream.